refactor: consolidate HR zone logic and remove redundancy
